demo-healthcare-nlp/cloud_function/requirements.txt
google-cloud-storage==2.5.0
google-cloud-bigquery==3.3.1
google-cloud-core==2.3.2
